fix: 피드백의 작성자 이거나 수신자만 삭제할 수 있는 기능

+++ backend/src/main/java/com/woowacourse/levellog/application/FeedbackService.java
@@ -13,6 +13,7 @@ import com.woowacourse.levellog.dto.FeedbackResponse;
 import com.woowacourse.levellog.dto.FeedbacksResponse;
 import com.woowacourse.levellog.dto.MemberResponse;
 import com.woowacourse.levellog.exception.FeedbackNotFoundException;
+import com.woowacourse.levellog.exception.InvalidFeedbackException;
 import com.woowacourse.levellog.exception.LevellogNotFoundException;
 import java.util.List;
 import java.util.stream.Collectors;
@@ -68,7 +69,15 @@ public class FeedbackService {
                 request.getFeedback().getEtc());
     }
 
-    public void deleteById(final Long id) {
+    public void deleteById(final Long id, final Long memberId) {
+        final Feedback feedback = feedbackRepository.findById(id)
+                .orElseThrow(FeedbackNotFoundException::new);
+
+        final Member member = memberRepository.findById(memberId)
+                .orElseThrow(MemberNotFoundException::new);
+        if (!feedback.isAssociatedWith(member)) {
+            throw new InvalidFeedbackException("자신이 남기거나 받은 피드백만 삭제할 수 있습니다.");
+        }
         feedbackRepository.deleteById(id);
     }
